ref(apm): Always use monotonic clock for time calculations
Also remove offset time calculations since everything is based on timeOrigin
